+++ RELEASE_NOTES.md
@@ -40,6 +40,7 @@
 - The font size of panel IDs in Framing Assistant has been increased to a legible level
 - The $$FRAMENR$$ placeholder is now saving the number as 4 digits
 - Framing - Rectangle panel number is now scaled based on rectangle size.
+- When loading an image from a file into framing the rectangle was not calculated correctly
 
 # Version 1.9